+++ src/searchPage/searchPageClient.ts
@@ -1,101 +1,12 @@
-import CoveoAnalyticsClient, { ClientOptions } from './analytics';
+import CoveoAnalyticsClient, { ClientOptions } from '../client/analytics';
 import { SearchEventRequest, ClickEventRequest, DocumentInformation } from '../events';
+import { SearchPageEvents, OmniboxSuggestionsMetadata, FacetMetadata, FacetRangeMetadata, CategoryFacetMetadata, DocumentIdentifier } from './searchPageEvents';
 
-export enum SearchPageEvents {
-    /**
-     * Identifies the search event that gets logged when the initial query is performed as a result of loading a search interface.
-     */
-    interfaceLoad = 'interfaceLoad',
-    /**
-     * Identifies the search event that gets logged when a new tab is selected in the search interface.
-     */
-    interfaceChange = 'interfaceChange',
-    /**
-     * Identifies the search event that gets logged when `enableAutoCorrection: true` and the query is automatically corrected.
-     */
-    didyoumeanAutomatic = 'didyoumeanAutomatic',
-    /**
-     * Identifies the search event that gets logged when the query suggestion with the corrected term is selected and successfully updates the query.
-     */
-    didyoumeanClick = 'didyoumeanClick',
-    /**
-     * Identifies the search event that gets logged when a sorting method is selected.
-     */
-    resultsSort = 'resultsSort',
-    /**
-     * Identifies the search event that gets logged when a submit button is selected on a search box.
-     */
-    searchboxSubmit = 'searchboxSubmit',
-    /**
-     * Identifies the search event that gets logged when a clear button is selected on a search box.
-     */
-    searchboxClear = 'searchboxClear',
-    /**
-     * The search-as-you-type event that gets logged when a query is automatically generated, and results are displayed while a user is entering text in the search box before they voluntarily submit the query.
-     */
-    searchboxAsYouType = 'searchboxAsYouType',
-    /**
-     * The event that gets logged when a breadcrumb facet is selected and the query is updated.
-     */
-    breadcrumbFacet = 'breadcrumbFacet',
-    /**
-     * Identifies the search event that gets logged when the event to clear the current breadcrumbs is triggered.
-     */
-    breadcrumbResetAll = 'breadcrumbResetAll',
-    /**
-     * Identifies the click event that gets logged when the Quick View element is selected and a Quick View modal of the document is displayed.
-     */
-    documentQuickview = 'documentQuickview',
-    /**
-     * Identifies the click event that gets logged when a user clicks on a search result to open an item.
-     */
-    documentOpen = 'documentOpen',
-    /**
-     * Identifies the search event that gets logged when a user clicks a query suggestion based on the usage analytics recorded queries.
-     */
-    omniboxAnalytics = 'omniboxAnalytics',
-    /**
-     * Identifies the search event that gets logged when a suggested search query is selected from a standalone searchbox.
-     */
-    omniboxFromLink = 'omniboxFromLink',
-}
 
-interface SearchPageClientProvider {
+export interface SearchPageClientProvider {
     getBaseMetadata: () => Record<string, any>;
     getSearchEventRequestPayload: () => Omit<SearchEventRequest, 'actionCause' | 'searchQueryUid'>
     getSearchUID: () => string;
-}
-
-export interface FacetMetadata {
-    facetId: string;
-    facetField: string;
-    facetValue: string;
-    facetTitle: string;
-}
-
-export interface FacetRangeMetadata extends Omit<FacetMetadata, 'facetValue'> {
-    facetRangeStart: string;
-    facetRangeEnd: string;
-    facetRangeEndInclusive: boolean;
-}
-
-export interface CategoryFacetMetadata {
-    categoryFacetId: string;
-    categoryFacetField: string;
-    categoryFacetPath: string[];
-    categoryFacetTitle: string;
-}
-
-export interface OmniboxSuggestionsMetadata {
-    suggestionRanking: number;
-    partialQueries: string;
-    suggestions: string;
-    partialQuery: string;
-}
-
-export interface DocumentIdentifier {
-    contentIDKey: string;
-    contentIDValue: string;
 }
 
 export class CoveoSearchPageClient {
